Revert wallet changes
