+++ wallet/src/account/mod.rs
@@ -18,13 +18,11 @@ mod output_cache;
 pub mod transaction_list;
 mod utxo_selector;
 
-mod partially_signed_transaction;
-pub use partially_signed_transaction::PartiallySignedTransaction;
-
 use common::address::pubkeyhash::PublicKeyHash;
 use common::chain::block::timestamp::BlockTimestamp;
 use common::chain::classic_multisig::ClassicMultisigChallenge;
-use common::chain::{AccountCommand, AccountOutPoint, AccountSpending};
+use common::chain::signature::sighash::signature_hash;
+use common::chain::{AccountCommand, AccountOutPoint, AccountSpending, TransactionCreationError};
 use common::primitives::id::WithId;
 use common::primitives::{Idable, H256};
 use common::size_estimation::{
@@ -52,6 +50,7 @@ use crate::wallet_events::{WalletEvents, WalletEventsNoOp};
 use crate::{get_tx_output_destination, SendRequest, WalletError, WalletResult};
 use common::address::{Address, RpcAddress};
 use common::chain::output_value::OutputValue;
+use common::chain::signature::inputsig::InputWitness;
 use common::chain::tokens::{
     make_token_id, IsTokenUnfreezable, NftIssuance, NftIssuanceV0, RPCFungibleTokenInfo, TokenId,
 };
@@ -65,6 +64,7 @@ use crypto::key::hdkd::u31::U31;
 use crypto::key::{PrivateKey, PublicKey};
 use crypto::vrf::VRFPublicKey;
 use itertools::{izip, Itertools};
+use serialization::{Decode, Encode};
 use std::cmp::Reverse;
 use std::collections::btree_map::Entry;
 use std::collections::{BTreeMap, BTreeSet};
@@ -104,6 +104,107 @@ impl TransactionToSign {
         match self {
             Self::Tx(tx) => HexEncoded::new(tx).to_string(),
             Self::Partial(tx) => HexEncoded::new(tx).to_string(),
+        }
+    }
+}
+
+#[derive(Debug, Eq, PartialEq, Clone, Encode, Decode)]
+pub struct PartiallySignedTransaction {
+    tx: Transaction,
+    witnesses: Vec<Option<InputWitness>>,
+
+    input_utxos: Vec<Option<TxOutput>>,
+    destinations: Vec<Option<Destination>>,
+}
+
+impl PartiallySignedTransaction {
+    pub fn new(
+        tx: Transaction,
+        witnesses: Vec<Option<InputWitness>>,
+        input_utxos: Vec<Option<TxOutput>>,
+        destinations: Vec<Option<Destination>>,
+    ) -> WalletResult<Self> {
+        ensure!(
+            tx.inputs().len() == witnesses.len(),
+            TransactionCreationError::InvalidWitnessCount
+        );
+
+        ensure!(
+            input_utxos.len() == witnesses.len(),
+            TransactionCreationError::InvalidWitnessCount
+        );
+
+        ensure!(
+            input_utxos.len() == destinations.len(),
+            TransactionCreationError::InvalidWitnessCount
+        );
+
+        Ok(Self {
+            tx,
+            witnesses,
+            input_utxos,
+            destinations,
+        })
+    }
+
+    pub fn new_witnesses(mut self, witnesses: Vec<Option<InputWitness>>) -> Self {
+        self.witnesses = witnesses;
+        self
+    }
+
+    pub fn tx(&self) -> &Transaction {
+        &self.tx
+    }
+
+    pub fn take_tx(self) -> Transaction {
+        self.tx
+    }
+
+    pub fn input_utxos(&self) -> &[Option<TxOutput>] {
+        self.input_utxos.as_ref()
+    }
+
+    pub fn destinations(&self) -> &[Option<Destination>] {
+        self.destinations.as_ref()
+    }
+
+    pub fn witnesses(&self) -> &[Option<InputWitness>] {
+        self.witnesses.as_ref()
+    }
+
+    pub fn count_inputs(&self) -> usize {
+        self.tx.inputs().len()
+    }
+
+    pub fn count_completed_signatures(&self) -> usize {
+        self.witnesses.iter().filter(|w| w.is_some()).count()
+    }
+
+    pub fn is_fully_signed(&self, chain_config: &ChainConfig) -> bool {
+        let inputs_utxos_refs: Vec<_> = self.input_utxos.iter().map(|out| out.as_ref()).collect();
+        self.witnesses
+            .iter()
+            .enumerate()
+            .zip(&self.destinations)
+            .all(|((input_num, w), d)| match (w, d) {
+                (Some(InputWitness::NoSignature(_)), None) => true,
+                (Some(InputWitness::NoSignature(_)), Some(_)) => false,
+                (Some(InputWitness::Standard(_)), None) => false,
+                (Some(InputWitness::Standard(sig)), Some(dest)) => {
+                    signature_hash(sig.sighash_type(), &self.tx, &inputs_utxos_refs, input_num)
+                        .and_then(|sighash| sig.verify_signature(chain_config, dest, &sighash))
+                        .is_ok()
+                }
+                (None, _) => false,
+            })
+    }
+
+    pub fn into_signed_tx(self, chain_config: &ChainConfig) -> WalletResult<SignedTransaction> {
+        if self.is_fully_signed(chain_config) {
+            let witnesses = self.witnesses.into_iter().map(|w| w.expect("cannot fail")).collect();
+            Ok(SignedTransaction::new(self.tx, witnesses)?)
+        } else {
+            Err(WalletError::FailedToConvertPartiallySignedTx(self))
         }
     }
 }
@@ -1266,7 +1367,7 @@ impl Account {
         Ok((
             txo.clone(),
             get_tx_output_destination(txo, &|pool_id| self.output_cache.pool_data(*pool_id).ok())
-                .ok_or(WalletError::InputCannotBeSpent(outpoint.clone()))?,
+                .ok_or(WalletError::InputCannotBeSpent(txo.clone()))?,
         ))
     }
